# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/update-email": {
            "post": {
                "tags": [
                    "Change Email"
                ],
                "summary": "User Send code for change",
                "description": "User Send code for change",
                "operationId": "User Send code for change",
                "requestBody": {
                    "description": "User Send code for change",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "random_int": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "123456"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "api/change-number": {
            "post": {
                "tags": [
                    "Change Number"
                ],
                "summary": "User Send Number for change",
                "description": "User Send Number for change",
                "operationId": "User Send Number for change",
                "requestBody": {
                    "description": "User Send Number for change",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "1234567\n     * "
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/update-number": {
            "post": {
                "tags": [
                    "Change Number"
                ],
                "summary": "User Send code for change",
                "description": "User Send code for change",
                "operationId": "User Send code ",
                "requestBody": {
                    "description": "User Send code for change",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "random_int": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "123456"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/registration": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "description": "User Register here",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "surname",
                                    "patronymic",
                                    "city",
                                    "username",
                                    "date_of_birth",
                                    "number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    },
                                    "surname": {
                                        "type": "text"
                                    },
                                    "patronymic": {
                                        "type": "text"
                                    },
                                    "city": {
                                        "type": "text"
                                    },
                                    "username": {
                                        "type": "text"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "number": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Register here",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}